Organise cypress tests better
