+++ cypress/e2e/case-list/header.cy.ts
@@ -1,4 +1,4 @@
-import { loginAndVisit } from "../support/helpers"
+import { loginAndVisit } from "../../support/helpers"
 
 describe("Home", () => {
   context("720p resolution", () => {
